+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v0.7.0 (TBD)
  - [Query] Fix for [Issue #37](https://github.com/cutright/DVH-Analytics/issues/37)
  - [Query] Proper min and max label updates with units in numerical query dialog
  - [RadBio] Fix for [Issue #34](https://github.com/cutright/DVH-Analytics/issues/34)
+ - [Time Series] Calculate tests for normality, two sample t-test, and Wilcoxon ranksum
  - [Misc] Corrected units for beam perimeter and area